+++ packages/superposition/src/Plan/AlivePlan.ts
@@ -1,23 +1,24 @@
-import { Kind, SyncAsync, UnaryFunction } from '@jamashita/anden-type';
+import { Kind, UnaryFunction } from '@jamashita/anden-type';
 import { MapPlan } from '@jamashita/genitore-plan';
 import { Chrono } from '../Chrono/Interface/Chrono';
+import { Bdb } from '../Interface/Bdb';
 import { Detoxicated } from '../Interface/Detoxicated';
 import { containsError, isSuperposition, ISuperposition } from '../Interface/ISuperposition';
 
 export class AlivePlan<A, B, E extends Error> implements MapPlan<Detoxicated<A>, 'AlivePlan'> {
   public readonly noun: 'AlivePlan' = 'AlivePlan';
-  private readonly mapper: UnaryFunction<Detoxicated<A>, SyncAsync<Detoxicated<B> | ISuperposition<B, E>>>;
+  private readonly mapper: UnaryFunction<Detoxicated<A>, PromiseLike<ISuperposition<B, E>> | ISuperposition<B, E> | PromiseLike<Bdb<B>> | Bdb<B>>;
   private readonly chrono: Chrono<B, E>;
 
   public static of<AT, BT, ET extends Error>(
-    mapper: UnaryFunction<Detoxicated<AT>, SyncAsync<Detoxicated<BT> | ISuperposition<BT, ET>>>,
+    mapper: UnaryFunction<Detoxicated<AT>, PromiseLike<ISuperposition<BT, ET>> | ISuperposition<BT, ET> | PromiseLike<Bdb<BT>> | Bdb<BT>>,
     chrono: Chrono<BT, ET>
   ): AlivePlan<AT, BT, ET> {
     return new AlivePlan<AT, BT, ET>(mapper, chrono);
   }
 
   protected constructor(
-    mapper: UnaryFunction<Detoxicated<A>, SyncAsync<Detoxicated<B> | ISuperposition<B, E>>>,
+    mapper: UnaryFunction<Detoxicated<A>, PromiseLike<ISuperposition<B, E>> | ISuperposition<B, E> | PromiseLike<Bdb<B>> | Bdb<B>>,
     chrono: Chrono<B, E>
   ) {
     this.mapper = mapper;
@@ -26,27 +27,27 @@ export class AlivePlan<A, B, E extends Error> implements MapPlan<Detoxicated<A>,
 
   public onMap(value: Detoxicated<A>): unknown {
     try {
-      const mapped: SyncAsync<Detoxicated<B> | ISuperposition<B, E>> = this.mapper(value);
+      const mapped: PromiseLike<ISuperposition<B, E>> | ISuperposition<B, E> | PromiseLike<Bdb<B>> | Bdb<B> = this.mapper(value);
 
-      if (isSuperposition<B, E>(mapped)) {
-        return this.forSuperposition(mapped);
-      }
-      if (Kind.isPromiseLike<Detoxicated<B> | ISuperposition<B, E>>(mapped)) {
+      if (Kind.isPromiseLike<Bdb<B> | ISuperposition<B, E>>(mapped)) {
         return mapped.then<unknown, unknown>(
-          (v: Detoxicated<B> | ISuperposition<B, E>) => {
+          (v: Bdb<B> | ISuperposition<B, E>) => {
             if (isSuperposition<B, E>(v)) {
               return this.forSuperposition(v);
             }
 
-            return this.chrono.accept(v);
+            return this.sync(v);
           },
           (e: unknown) => {
             return this.forError(e);
           }
         );
       }
+      if (isSuperposition<B, E>(mapped)) {
+        return this.forSuperposition(mapped);
+      }
 
-      return this.chrono.accept(mapped);
+      return this.sync(mapped);
     }
     catch (err: unknown) {
       return this.forError(err);
@@ -67,6 +68,14 @@ export class AlivePlan<A, B, E extends Error> implements MapPlan<Detoxicated<A>,
         return this.chrono.throw(c);
       }
     );
+  }
+
+  private sync(v: Bdb<B>): unknown {
+    if (v instanceof Error) {
+      return this.forError(v);
+    }
+
+    return this.chrono.accept(v);
   }
 
   private forError(e: unknown): unknown {