+++ packages/unscharferelation/src/Heisenberg/HeisenbergType.ts
@@ -1,0 +1,1 @@
+export type HeisenbergType = 'Absent' | 'Lost' | 'Present' | 'Uncertain';