+++ packages/unscharferelation/src/Interface/Ymy.ts
@@ -1,0 +1,4 @@
+import { Suspicious } from '@jamashita/anden-type';
+import { Matter } from './Matter';
+
+export type Ymy<T> = Suspicious<Matter<T>>;